+++ app/controllers/rails_chatbot/chat_controller.rb
@@ -2,6 +2,11 @@ module RailsChatbot
   class ChatController < ApplicationController
     def index
       # Main chat interface
+      @rails_chatbot_routes = {
+        conversations_path: rails_chatbot.conversations_path,
+        conversation_messages_path_template: rails_chatbot.conversation_messages_path(':id'),
+        messages_path: rails_chatbot.messages_path
+      }.to_json
     end
 
     def search